Proformas module refactored
