+++ apps/frontend/src/app/modules/proformas/proformas.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { Routes, RouterModule } from '@angular/router';
 import { ProformasComponent } from './proformas.component';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { ReactiveFormsModule } from '@angular/forms';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatButtonModule } from '@angular/material/button';
@@ -11,20 +11,23 @@ import { MatNativeDateModule } from '@angular/material/core';
 import { MatIconModule } from '@angular/material/icon';
 import { MatSelectModule } from '@angular/material/select';
 import { MatCheckboxModule } from '@angular/material/checkbox';
-import { MatGridListModule } from '@angular/material/grid-list';
 import { MatCardModule } from '@angular/material/card';
 import { MatRadioModule } from '@angular/material/radio';
 import { MatTableModule } from '@angular/material/table';
-
 import { MatTooltipModule } from '@angular/material/tooltip';
+import { MatPaginatorModule } from '@angular/material/paginator';
+import { ProformaComponent } from './proforma/proforma.component';
 
-const routes: Routes = [{ path: '', component: ProformasComponent }];
+const routes: Routes = [
+  { path: '', component: ProformasComponent },
+  { path: 'new', component: ProformaComponent },
+  { path: ':id', component: ProformaComponent },
+];
 
 @NgModule({
-  declarations: [ProformasComponent],
+  declarations: [ProformasComponent, ProformaComponent],
   imports: [
     CommonModule,
-    FormsModule,
     ReactiveFormsModule,
     MatFormFieldModule,
     MatInputModule,
@@ -34,13 +37,12 @@ const routes: Routes = [{ path: '', component: ProformasComponent }];
     MatIconModule,
     MatSelectModule,
     MatCheckboxModule,
-    MatGridListModule,
     MatCardModule,
     MatRadioModule,
     MatTableModule,
     MatTooltipModule,
-
     RouterModule.forChild(routes),
+    MatPaginatorModule,
   ],
 })
 export class ProformasModule {}
